Remove dead code, deduplicate getNestedValue, simplify types
Dead code removed (replaced by workflow steps):
- case_matching/case_matcher.ts + entity_index.ts (replaced by case_matching_step)
- risk_scoring/entity_risk_enrichment.ts (was only used by removed API routes)
- 19 tests for dead modules

Refactored:
- Extract getNestedValue to shared utils/get_nested_value.ts (was duplicated
  in feature_extraction.ts and extract_entities.ts)
- Simplify types.ts: remove PipelineConfig, PipelineExecutionResult,
  ProcessedAlertTracker, IncrementalAdConfig, CaseMatchScore, CaseMatchingConfig,
  EntityWeights, DeduplicationConfig (all from old orchestrator design)
- Keep only: EntityExtractionConfig, ObservableTypeKey, ExtractedEntity,
  DEFAULT_ENTITY_EXTRACTION_CONFIG

All 88 tests passing (61 pipeline + 22 inline tools + 5 cases), 0 type errors.
