+++ x-pack/solutions/security/plugins/elastic_assistant/server/index.ts
@@ -28,9 +28,4 @@ export type {
 // Alert Investigation Pipeline - pure algorithm exports for Agent Builder tools
 export { deduplicateAlerts } from './lib/alert_investigation/deduplication';
 export { extractEntitiesFromAlerts } from './lib/alert_investigation/entity_extraction';
-export { matchAlertsToCases } from './lib/alert_investigation/case_matching';
-export type {
-  ExtractedEntity,
-  CaseMatchScore,
-  ObservableTypeKey,
-} from './lib/alert_investigation/types';
+export type { ExtractedEntity, ObservableTypeKey } from './lib/alert_investigation/types';
