--- x-pack/solutions/security/plugins/elastic_assistant/server/lib/alert_investigation/case_matching/index.ts
@@ -1,9 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-export { matchAlertsToCases } from './case_matcher';
-export type { CaseMatchResult, CaseMatchingResult } from './case_matcher';
