+++ x-pack/solutions/security/plugins/elastic_assistant/server/lib/alert_investigation/index.ts
@@ -8,10 +8,5 @@
 export { registerPipelineWorkflowSteps } from './workflow_steps';
 export { deduplicateAlerts } from './deduplication';
 export { extractEntitiesFromAlerts } from './entity_extraction';
-export { matchAlertsToCases } from './case_matching';
 export { WorkflowInitService } from './workflow_init';
-export type {
-  ExtractedEntity,
-  CaseMatchScore,
-  ObservableTypeKey,
-} from './types';
+export type { ExtractedEntity, ObservableTypeKey } from './types';
