--- x-pack/solutions/security/plugins/elastic_assistant/server/lib/alert_investigation/risk_scoring/index.ts
@@ -1,9 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-export { enrichAlertsWithEntityRisk } from './entity_risk_enrichment';
-export type { EntityRiskScore, AlertWithRiskScore } from './entity_risk_enrichment';
